Use lib.rs to define cli config
Trying to follow those rust notes from the book about lib.rs to contain
more of the program and to try to make the main.rs file small and such.

The trick is that I'm not sure how to like, move all the tokio stuff
away?

What I'd like to do is setup two event loops here, one to watch OBS for
scene changes so that I can fetch the image source to update from the
current scene, and then igure out a nicer way to have the events from
OBS go into a queue and build up a context string that I trim down for a
bit.