Move OBS functionality to own file + struct
Still could be cleaned up a bit more, like moving the async private
methods of the module into the struct itself, but for now this is fine.

Eventually I'd like to make a trait for a general AvatarController or
something similar, but that can wait until I decide to use egui or
nannou for all this stuff.