Tweak rules based on local testing
No where near sensible, but they kind of sort of work sometimes and
that's helpful that it occasionally hits the right face to show when I
say specific things or when the sentimentality moon is in the right
place.

I won't lie though, while it's interesting and seemed like a proper use
for the sentiment analysis tool, I somewhat think a finite state machine
might still be a better option overall and have it get some triggers and
rules setup for that instead. Specifically, I was amusing myself by
making "blank" faces for some of the emotions and whatnot and it would
probably be better if that blankness was a state or something overall
and then the other bits of the equation would select the potential
options and emotions to show while the eyes are blank.