Basic duplication detection (ephemereal)
Able to detect a new file being added that is a duplicate of an old one.
Note that the debounced event doesn't separate our renames so it falsely
claims there's a duplicate from that.

We're not persisting the DB down yet, so it's just checking in on it in
memory