Use debounce-full to get item types +fix dupcheck
debounce mini doesn't provide the types of changes which makes handling
renames weird and awkward. Not to mention we don't care about access
events since that's not changing anything.

Also the debounce full has the property that a rename sends the paths as
from then to in that order, so when we process it we remove the old one
first then add the new one, so no false positive dup is detected like
what was happening with the mini version.