Intentionally Poison for testing purposes
This 'works' but is it actually what one should do? The sender channel
becomes kaput on a panic in one of the workers, which feels very much
like an isolation problem. And so, we have to create entirely new
workers, so like, what's going to happen to all the other ones? Is there
a good way to observe and view if this is going to cause problems or
result in workers dangling around in memory that aren't tracked or will
they be properly dropped by the replacement?